splits models out into individual modules
