+++ src/models/block.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct Height {
+    /// The current block height of the chain.
+    pub height: u64,
+}
