+++ src/models/geocode.rs
@@ -1,0 +1,21 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct Geocode {
+    /// The long version of city for the last asserted location
+    pub long_city: Option<String>,
+    /// The long version of country for the last asserted location
+    pub long_country: Option<String>,
+    /// The long version of state for the last asserted location
+    pub long_state: Option<String>,
+    /// The long version of street for the last asserted location
+    pub long_street: Option<String>,
+    /// The short version of city for the last asserted location
+    pub short_city: Option<String>,
+    /// The short version of country for the last asserted location
+    pub short_country: Option<String>,
+    /// The short version of state for the last asserted location
+    pub short_state: Option<String>,
+    /// The short version of street for the last asserted location
+    pub short_street: Option<String>,
+}
