+++ src/models/transaction.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Clone, Deserialize, Debug)]
+pub struct PendingTxnStatus {
+    pub hash: String,
+}
