+++ src/models/values.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use crate::{Error, Result};
 use core::fmt;
 use rust_decimal::prelude::*;
 use serde::{de::Deserializer, Deserialize, Serialize};
